requirements.txt
# Path: requirements.txt
pandas
numpy
PyYAML
requests
python-dotenv
databento
twelvedata
pyarrow

# Visualisation (spiders / reporting)
plotly
kaleido